AvailablePharmacyRepository primary key changed from int to long

+++ src/main/java/com/dstym/pharmaciesondutyattica/repository/AvailablePharmacyRepository.java
@@ -3,6 +3,6 @@ package com.dstym.pharmaciesondutyattica.repository;
 import com.dstym.pharmaciesondutyattica.entity.AvailablePharmacy;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface AvailablePharmacyRepository extends JpaRepository<AvailablePharmacy, Integer> {
+public interface AvailablePharmacyRepository extends JpaRepository<AvailablePharmacy, Long> {
 
 }
